backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
pyyaml
python-dotenv
pytest
pytest-asyncio
respx
